remove details page!
